Fix No picture being returned from API.
I used a series of breakpoint to find the error in the code. Used print statements along the way to get feedback as to what was working.
Eventually found that API Key was invalid and server was return a 100 status code.
Corrected typo in API Key and re-launched app.

App works as expected now.